bugfix: activate the initialization fence unless it's been explicitly SKIPPED
The problem was that this hook exits if $INITHOOKS_CONF already exists.  So the
initialization fence isn't activated. But you can't trigger turnkey-init on
first login if you can't login as root so you need to create that file in order
to set the root password.

In other words 29preseed assumes preseeding is all or nothing, which nullifies
the value of the initialization fence, which is intended to make initialization
easy on platforms that won't implement full integration with TurnKey's
preseeding mechanism (E.g., like the TurnKey Hub)